% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme4.R
\name{pipe_lme4}
\alias{pipe_lme4}
\alias{lmer}
\alias{lmer.default}
\alias{lmer.data.frame}
\alias{glmer}
\alias{glmer.default}
\alias{glmer.data.frame}
\title{Pipe-friendly wrappers for mixed-effects models}
\usage{
lmer(x, ...)

\method{lmer}{default}(x, ...)

\method{lmer}{data.frame}(x, formula, ...)

glmer(x, ...)

\method{glmer}{default}(x, ...)

\method{glmer}{data.frame}(x, formula, ...)
}
\arguments{
\item{x}{For \code{data.frame} methods: a data frame to use as the \code{data}
argument. For \code{default} methods: passed straight through to the
corresponding \verb{lme4::} function.}

\item{...}{Additional arguments passed to the underlying function.}

\item{formula}{A mixed-effects model formula, passed to \code{\link[lme4:glmer]{lme4::glmer()}}.}
}
\value{
For \code{lmer}: an object of class \code{"lmerMod"}, as returned by
\code{\link[lme4:lmer]{lme4::lmer()}}.

For \code{glmer}: an object of class \code{"glmerMod"}, as returned by
\code{\link[lme4:glmer]{lme4::glmer()}}.
}
\description{
These functions wrap \code{\link[lme4:lmer]{lme4::lmer()}} and \code{\link[lme4:glmer]{lme4::glmer()}} as proper S3
generics, adding \code{data.frame} methods so that a data frame can be passed as
the first argument via the base pipe (\verb{|>}).
}
\details{
These functions require the \code{lme4} package to be installed.
}
\examples{
if (requireNamespace("lme4", quietly = TRUE)) {
  lme4::sleepstudy |> lmer(Reaction ~ Days + (Days | Subject))
}

if (requireNamespace("lme4", quietly = TRUE)) {
  lme4::cbpp |>
    glmer(cbind(incidence, size - incidence) ~ period + (1 | herd),
          family = binomial)
}

}
