% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASS.R
\name{pipe_MASS}
\alias{pipe_MASS}
\alias{glm_nb}
\alias{glm_nb.default}
\alias{glm_nb.data.frame}
\alias{polr}
\alias{polr.default}
\alias{polr.data.frame}
\alias{lda}
\alias{lda.default}
\alias{lda.data.frame}
\alias{qda}
\alias{qda.default}
\alias{qda.data.frame}
\title{Pipe-friendly wrappers for MASS modelling functions}
\usage{
glm_nb(x, ...)

\method{glm_nb}{default}(x, ...)

\method{glm_nb}{data.frame}(x, formula, ...)

polr(x, ...)

\method{polr}{default}(x, ...)

\method{polr}{data.frame}(x, formula, ...)

lda(x, ...)

\method{lda}{default}(x, ...)

\method{lda}{data.frame}(x, formula, ...)

qda(x, ...)

\method{qda}{default}(x, ...)

\method{qda}{data.frame}(x, formula, ...)
}
\arguments{
\item{x}{For \code{data.frame} methods: a data frame to use as the \code{data}
argument. For \code{default} methods: passed straight through to the
corresponding \verb{MASS::} function.}

\item{...}{Additional arguments passed to the underlying function.}

\item{formula}{A formula, passed to \code{\link[MASS:qda]{MASS::qda()}}.}
}
\value{
For \code{glm_nb} / \code{glm.nb}: an object of class
\code{c("negbin", "glm", "lm")}, as returned by \code{\link[MASS:glm.nb]{MASS::glm.nb()}}.

For \code{polr}: an object of class \code{"polr"}, as returned by
\code{\link[MASS:polr]{MASS::polr()}}.

For \code{lda}: an object of class \code{"lda"}, as returned by \code{\link[MASS:lda]{MASS::lda()}}.

For \code{qda}: an object of class \code{"qda"}, as returned by \code{\link[MASS:qda]{MASS::qda()}}.
}
\description{
These functions wrap \code{\link[MASS:glm.nb]{MASS::glm.nb()}}, \code{\link[MASS:polr]{MASS::polr()}}, \code{\link[MASS:lda]{MASS::lda()}}, and
\code{\link[MASS:qda]{MASS::qda()}} as proper S3 generics, adding \code{data.frame} methods so that a
data frame can be passed as the first argument via the base pipe (\verb{|>}).
}
\details{
These functions require the \code{MASS} package to be installed.
}
\examples{
if (requireNamespace("MASS", quietly = TRUE)) {
  MASS::quine |> glm_nb(Days ~ Sex + Age + Eth)
}

if (requireNamespace("MASS", quietly = TRUE)) {
  MASS::housing |> polr(Sat ~ Infl + Type + Cont)
}

if (requireNamespace("MASS", quietly = TRUE)) {
  iris |> lda(Species ~ .)
}

if (requireNamespace("MASS", quietly = TRUE)) {
  iris |> qda(Species ~ .)
}

}
