% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{pipe_survival}
\alias{pipe_survival}
\alias{survfit}
\alias{survfit.default}
\alias{survfit.data.frame}
\alias{coxph}
\alias{coxph.default}
\alias{coxph.data.frame}
\alias{survreg}
\alias{survreg.default}
\alias{survreg.data.frame}
\title{Pipe-friendly wrappers for survival analysis functions}
\usage{
survfit(x, ...)

\method{survfit}{default}(x, ...)

\method{survfit}{data.frame}(x, formula, ...)

coxph(x, ...)

\method{coxph}{default}(x, ...)

\method{coxph}{data.frame}(x, formula, ...)

survreg(x, ...)

\method{survreg}{default}(x, ...)

\method{survreg}{data.frame}(x, formula, ...)
}
\arguments{
\item{x}{For \code{data.frame} methods: a data frame to use as the \code{data}
argument. For \code{default} methods: passed straight through to the
corresponding \verb{survival::} function.}

\item{...}{Additional arguments passed to the underlying function.}

\item{formula}{A formula with a \code{\link[survival:Surv]{survival::Surv()}} object on the left-hand
side, passed to \code{\link[survival:survreg]{survival::survreg()}}.}
}
\value{
For \code{survfit}: an object of class \code{"survfit"}, as returned by
\code{\link[survival:survfit]{survival::survfit()}}.

For \code{coxph}: an object of class \code{"coxph"}, as returned by
\code{\link[survival:coxph]{survival::coxph()}}.

For \code{survreg}: an object of class \code{"survreg"}, as returned by
\code{\link[survival:survreg]{survival::survreg()}}.
}
\description{
These functions wrap \code{\link[survival:survfit]{survival::survfit()}}, \code{\link[survival:coxph]{survival::coxph()}}, and
\code{\link[survival:survreg]{survival::survreg()}} as proper S3 generics, adding \code{data.frame} methods so
that a data frame can be passed as the first argument via the base pipe
(\verb{|>}).
}
\details{
These functions require the \code{survival} package to be installed.
}
\examples{
if (requireNamespace("survival", quietly = TRUE)) {
  survival::lung |> survfit(survival::Surv(time, status) ~ sex)
}

if (requireNamespace("survival", quietly = TRUE)) {
  survival::lung |> coxph(survival::Surv(time, status) ~ age + sex)
}

if (requireNamespace("survival", quietly = TRUE)) {
  survival::lung |>
    survreg(survival::Surv(time, status) ~ age + sex, dist = "weibull")
}

}
