% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{pipe_tests}
\alias{pipe_tests}
\alias{t.test.data.frame}
\alias{wilcox.test.data.frame}
\title{Pipe-friendly data.frame methods for t.test and wilcox.test}
\usage{
\method{t.test}{data.frame}(x, formula, ...)

\method{wilcox.test}{data.frame}(x, formula, ...)
}
\arguments{
\item{x}{A data frame.}

\item{formula}{A formula of the form \code{lhs ~ rhs}.}

\item{...}{Additional arguments passed to the underlying stats function.}
}
\value{
For \code{t.test}: an object of class \code{"htest"}, as returned by
\code{\link[stats:t.test]{stats::t.test()}}.

For \code{wilcox.test}: an object of class \code{"htest"}, as returned by
\code{\link[stats:wilcox.test]{stats::wilcox.test()}}.
}
\description{
S3 \code{data.frame} methods for \code{\link[stats:t.test]{stats::t.test()}} and \code{\link[stats:wilcox.test]{stats::wilcox.test()}}
that allow a data frame to be passed as the first argument via the base pipe
(\verb{|>}), avoiding the need for the \verb{data = _} placeholder syntax.
}
\examples{
# Without pipedreams:
# mtcars |> t.test(mpg ~ am, data = _)

# With pipedreams:
mtcars |> t.test(mpg ~ am)

mtcars |> wilcox.test(mpg ~ am)

}
