% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipedreams-package.R
\docType{package}
\name{pipedreams-package}
\alias{pipedreams}
\alias{pipedreams-package}
\title{pipedreams: Pipe-Friendly Wrappers for Common Statistical Tests and Models}
\description{
Provides \code{data.frame} S3 methods for a range of statistical functions so
that a data frame can be passed as the first argument via the base pipe
(\verb{|>}) without needing the \verb{data = _} placeholder.
}
\details{
Covered functions:
\itemize{
\item \strong{stats}: \code{t.test}, \code{wilcox.test}, \code{lm}, \code{glm}, \code{aov}, \code{nls}
\item \strong{survival}: \code{survfit}, \code{coxph}, \code{survreg}
\item \strong{lme4}: \code{lmer}, \code{glmer}
\item \strong{MASS}: \code{glm.nb}, \code{polr}, \code{lda}, \code{qda}
}
}
\section{Usage}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(pipedreams)

# base stats
mtcars |> t.test(mpg ~ am)
mtcars |> wilcox.test(mpg ~ am)
mtcars |> lm(mpg ~ wt + cyl)
mtcars |> glm(am ~ mpg, family = binomial)
npk    |> aov(yield ~ block + N * P * K)
Puromycin |> subset(state == "treated") |>
  nls(rate ~ Vm * conc / (K + conc), start = list(Vm = 200, K = 0.05))

# survival
survival::lung |> survfit(Surv(time, status) ~ sex)
survival::lung |> coxph(Surv(time, status) ~ age + sex)
survival::lung |> survreg(Surv(time, status) ~ age + sex, dist = "weibull")

# lme4
lme4::sleepstudy |> lmer(Reaction ~ Days + (Days | Subject))
lme4::cbpp |> glmer(cbind(incidence, size - incidence) ~ period + (1 | herd),
              family = binomial)

# MASS
MASS::quine   |> glm_nb(Days ~ Sex + Age + Eth)
MASS::housing |> polr(Sat ~ Infl + Type + Cont)
iris    |> lda(Species ~ .)
iris    |> qda(Species ~ .)
}\if{html}{\out{</div>}}
}

\author{
\strong{Maintainer}: Thulin Måns \email{mans@statistikkonsult.com}

}
\keyword{internal}
